Move app.js from components to pages
