+++ app/style.less
@@ -14,8 +14,6 @@
  */
 
 @import "core/core.less";
-
-@import "components/app/app.less";
 @import "components/navbar/navbar.less";
 @import "components/notification/notification.less";
 @import "components/inputgroup/inputgroup.less";
@@ -37,6 +35,7 @@
 @import "components/chart/settings.less";
 @import "components/chart/modalsubnav.less";
 
+@import "pages/app/app.less";
 @import "pages/login/login.less";
 @import "pages/signup/signup.less";
 @import "pages/profile/profile.less";
